Rename package to metrics to match folder path
