+++ pkg/otlp/metrics/summary_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2022-present Datadog, Inc.
 
-package translator
+package metrics
 
 import (
 	"testing"
